opencl: rename backend `opencl2` -> `opencl`
