+++ ggml/src/CMakeLists.txt
@@ -298,13 +298,6 @@ else ()
     ggml_add_cpu_backend_variant_impl("")
 endif()
 
-if (GGML_OPENCL)
-    set(GGML_OPENCL2 ON)
-    add_compile_definitions(GGML_USE_OPENCL)
-else ()
-    set(GGML_OPENCL2 OFF)
-endif ()
-
 ggml_add_backend(BLAS)
 ggml_add_backend(CANN)
 ggml_add_backend(CUDA)
@@ -315,7 +308,7 @@ ggml_add_backend(MUSA)
 ggml_add_backend(RPC)
 ggml_add_backend(SYCL)
 ggml_add_backend(Vulkan)
-ggml_add_backend(OpenCL2)
+ggml_add_backend(OpenCL)
 
 foreach (target ggml-base ggml)
     target_include_directories(${target} PUBLIC    $<BUILD_INTERFACE:${CMAKE_CURRENT_SOURCE_DIR}/../include> $<INSTALL_INTERFACE:include>)
