+++ ggml/src/ggml-opencl/kernels/embed_kernel.py
