+++ ggml/src/ggml-opencl/kernels/ggml-opencl2_mul_mat_Ab_Bi_8x4.cl
